Refactored component to simplify its behavior.
