+++ src/components/ApplicantsBadge.md
@@ -1,0 +1,11 @@
+
+ApplicantsBadge example:
+
+```js
+<ApplicantsBadge applicant={
+  {firstName: 'Applicant',
+   lastName: 'X-Team',
+   email: 'applicant@x-team.com',
+  }
+} />
+```
