--- src/components/ResultListItem.md
@@ -1,20 +1,0 @@
-
-ResultListItem example:
-
-```js
-<ResultListItem applicant={null} />
-```
-
-```js
-<ResultListItem applicant={
-  {firstName: 'Applicant',
-   lastName: 'X-Team',
-   email: 'applicant@x-team.com',
-   baseId: 'e303e627d9c3f8c47e3bff5',
-   tags: [
-     {name: 'react', portfolios: [{id: 1, body: '2 years of experience'}]},
-     {name: 'node.js'}
-   ]
-  }
-} />
-```
