+++ src/index.js
@@ -2,12 +2,12 @@ import ErrorBox from './components/ErrorBox'
 import PageTitle from './components/PageTitle'
 import XIcon from './components/XIcon'
 import CollapsibleSection from './components/CollapsibleSection'
-import ResultListItem from './components/ResultListItem'
+import ApplicantsBadge from './components/ApplicantsBadge'
 
 export {
   ErrorBox,
   PageTitle,
   XIcon,
   CollapsibleSection,
-  ResultLIstItem
+  ApplicantsBadge
 }
